# openapi.yml
openapi: 3.0.3
info:
  title: Gimpel API
  version: 1.0.0
  description: REST API for Gimpel module and deployment management

servers:
  - url: http://localhost:8080
    description: Local server

paths:
  /api/v1/modules:
    post:
      summary: Upload module
      operationId: uploadModule
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - id
                - version
                - signature
                - image
              properties:
                id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                version:
                  type: string
                protocol:
                  type: string
                signature:
                  type: string
                  description: Hex-encoded signature
                signed_at:
                  type: integer
                  format: int64
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Module uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadModuleResponse'
        '400':
          description: Invalid request
        '500':
          description: Server error

    get:
      summary: List modules
      operationId: listModules
      responses:
        '200':
          description: List of modules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListModulesResponse'
        '500':
          description: Server error

  /api/v1/modules/{id}/{version}:
    get:
      summary: Get module
      operationId: getModule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Module details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleInfo'
        '404':
          description: Module not found
        '500':
          description: Server error

    delete:
      summary: Delete module
      operationId: deleteModule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Module deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '500':
          description: Server error

  /api/v1/modules/{id}/{version}/download:
    get:
      summary: Download module image
      operationId: downloadModule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Module image file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Module not found

  /api/v1/satellites:
    get:
      summary: List satellites
      operationId: listSatellites
      responses:
        '200':
          description: List of satellites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSatellitesResponse'
        '500':
          description: Server error

  /api/v1/satellites/{id}:
    get:
      summary: Get satellite
      operationId: getSatellite
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Satellite details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SatelliteInfo'
        '404':
          description: Satellite not found
        '500':
          description: Server error

  /api/v1/satellites/{id}/deployments:
    post:
      summary: Create deployment
      operationId: createDeployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeploymentRequest'
      responses:
        '200':
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
        '400':
          description: Invalid request
        '404':
          description: Satellite not found
        '500':
          description: Server error

    get:
      summary: Get deployment
      operationId: getDeployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
        '404':
          description: Deployment not found
        '500':
          description: Server error

    delete:
      summary: Delete deployment
      operationId: deleteDeployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deployment deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '500':
          description: Server error

  /api/v1/deployments:
    get:
      summary: List deployments
      operationId: listDeployments
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployments:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeploymentResponse'
        '500':
          description: Server error

  /api/v1/pairings:
    post:
      summary: Create pairing token
      operationId: createPairing
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePairingRequest'
      responses:
        '200':
          description: Pairing created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PairingResponse'
        '500':
          description: Server error

components:
  schemas:
    UploadModuleResponse:
      type: object
      properties:
        id:
          type: string
        version:
          type: string
        digest:
          type: string
        signature:
          type: string
        signed_by:
          type: string
        signed_at:
          type: integer
          format: int64
        size:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time

    ListModulesResponse:
      type: object
      properties:
        modules:
          type: array
          items:
            $ref: '#/components/schemas/ModuleInfo'

    ModuleInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        version:
          type: string
        protocol:
          type: string
        digest:
          type: string
        size_bytes:
          type: integer
          format: int64
        signed_by:
          type: string
        signed_at:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time

    CreateDeploymentRequest:
      type: object
      properties:
        modules:
          type: array
          items:
            $ref: '#/components/schemas/ModuleAssignment'

    ModuleAssignment:
      type: object
      properties:
        module_id:
          type: string
        module_version:
          type: string
        execution_mode:
          type: string
        listeners:
          type: array
          items:
            $ref: '#/components/schemas/Listener'
        env:
          type: object
          additionalProperties:
            type: string

    Listener:
      type: object
      properties:
        id:
          type: string
        protocol:
          type: string
        port:
          type: integer
          format: uint32
        high_interaction:
          type: boolean

    DeploymentResponse:
      type: object
      properties:
        satellite_id:
          type: string
        version:
          type: integer
          format: int64
        modules:
          type: array
          items:
            $ref: '#/components/schemas/ModuleAssignment'
        updated_at:
          type: string
          format: date-time

    ListSatellitesResponse:
      type: object
      properties:
        satellites:
          type: array
          items:
            $ref: '#/components/schemas/SatelliteInfo'

    SatelliteInfo:
      type: object
      properties:
        id:
          type: string
        hostname:
          type: string
        ip_address:
          type: string
        os:
          type: string
        arch:
          type: string
        status:
          type: string
        registered_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time

    CreatePairingRequest:
      type: object
      properties:
        ttl_seconds:
          type: integer
          format: int64

    PairingResponse:
      type: object
      properties:
        id:
          type: string
        token:
          type: string
        expires_at:
          type: string
          format: date-time